fix: Resolve dynamic route conflict for customer APIs
I consolidated your customer API endpoints to use `[customerId]` consistently.

I removed `pages/api/admin/customers/[customerId].ts` (which was an older, less complete version).
I renamed `pages/api/admin/customers/[id].ts` to `pages/api/admin/customers/[customerId].ts`.
I also renamed the corresponding test file `pages/api/admin/customers/[id].test.ts` to `pages/api/admin/customers/[customerId].test.ts`.

This change standardizes the dynamic route parameter for customer-specific API endpoints, resolving a Next.js error: "You cannot use different slug names for the same dynamic path ('customerId' !== 'id')."
The more comprehensive and secure version of the GET/PUT handler for single customer operations is now correctly named and active.
